refactor: 集計関連ファイルをsrc/lib/agg/サブディレクトリに移動
aggregate, gtAggregator, crossAggregator, sqlHelpers, pivot, download を
src/lib/agg/ に集約し、全importパスを更新。
