+++ src/components/rightPane/AIBubble.ts
@@ -1,4 +1,4 @@
-import type { AggResult } from "../../lib/aggregate";
+import type { AggResult } from "../../lib/agg/aggregate";
 import type { LayoutMeta } from "../../lib/layout";
 import { isAIAvailable, generateComment } from "../../lib/aiComment";
 import { t } from "../../lib/i18n";
