+++ src/components/rightPane/ChartRenderer.ts
@@ -1,8 +1,8 @@
 /** Chart rendering component */
 
-import type { AggResult, QuestionDef } from "../../lib/aggregate";
+import type { AggResult, QuestionDef } from "../../lib/agg/aggregate";
 import type { LayoutMeta } from "../../lib/layout";
-import { pivot } from "../../lib/pivot";
+import { pivot } from "../../lib/agg/pivot";
 import { Chart, getSeriesColor, getThemeColors } from "../../lib/chartConfig";
 import { resolveQuestionLabel, resolveValueLabel, resolveSubLabel } from "../../lib/labelResolver";
 import { escHtml } from "../shared/escHtml";
