+++ src/lib/agg/download.ts
@@ -1,6 +1,6 @@
 import type { AggResult } from "./aggregate";
 import { parseCrossSub } from "./aggregate";
-import type { LayoutMeta } from "./layout";
+import type { LayoutMeta } from "../layout";
 import { NA_VALUE } from "./sqlHelpers";
 import { pivot } from "./pivot";
